requirements.txt
opencv-python
face_recognition
numpy
Pillow
Adafruit_DHT
RPi.GPIO
